+++ src/main/kotlin/me/nobaboy/nobaaddons/config/categories/FishingCategory.kt
@@ -11,7 +11,7 @@ import me.nobaboy.nobaaddons.features.fishing.crimsonisle.TrophyFishChat
 import me.nobaboy.nobaaddons.ui.TextShadow
 import me.nobaboy.nobaaddons.utils.CommonText
 import me.nobaboy.nobaaddons.utils.mc.TextUtils.white
-import me.nobaboy.nobaaddons.utils.enums.AnnounceChannel
+import me.nobaboy.nobaaddons.utils.hypixel.AnnounceChannel
 import me.nobaboy.nobaaddons.utils.tr
 import net.minecraft.text.Text
