+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/chatcommands/impl/party/TransferCommand.kt
@@ -4,7 +4,7 @@ import me.nobaboy.nobaaddons.api.PartyAPI
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatCommand
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatContext
 import me.nobaboy.nobaaddons.utils.mc.MCUtils
-import me.nobaboy.nobaaddons.utils.chat.HypixelCommands
+import me.nobaboy.nobaaddons.utils.hypixel.HypixelCommands
 
 class TransferCommand : ChatCommand() {
 	override val enabled: Boolean get() = config.party.transfer
