+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/chatcommands/impl/party/WarpCommand.kt
@@ -4,7 +4,7 @@ import me.nobaboy.nobaaddons.api.PartyAPI
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatCommand
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatContext
 import me.nobaboy.nobaaddons.utils.Scheduler
-import me.nobaboy.nobaaddons.utils.chat.HypixelCommands
+import me.nobaboy.nobaaddons.utils.hypixel.HypixelCommands
 import org.apache.commons.lang3.StringUtils
 import kotlin.time.Duration.Companion.seconds
