+++ src/main/kotlin/me/nobaboy/nobaaddons/features/chat/chatcommands/impl/shared/WarpOutCommand.kt
@@ -2,7 +2,7 @@ package me.nobaboy.nobaaddons.features.chat.chatcommands.impl.shared
 
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatCommand
 import me.nobaboy.nobaaddons.features.chat.chatcommands.ChatContext
-import me.nobaboy.nobaaddons.utils.chat.ChatUtils
+import me.nobaboy.nobaaddons.utils.mc.chat.ChatUtils
 import kotlin.time.Duration.Companion.seconds
 
 class WarpOutCommand(private var command: String, private val categoryEnabled: () -> Boolean) : ChatCommand(3.seconds) {
